+++ src/main/java/svenhjol/charmony/relics/common/CommonInitializer.java
@@ -4,7 +4,6 @@ import net.fabricmc.api.ModInitializer;
 import svenhjol.charmony.api.core.Side;
 import svenhjol.charmony.relics.RelicsMod;
 import svenhjol.charmony.relics.common.features.derelicts.Derelicts;
-import svenhjol.charmony.relics.common.features.elixirs.Elixirs;
 import svenhjol.charmony.relics.common.features.relics.Relics;
 
 public final class CommonInitializer implements ModInitializer {
@@ -16,7 +15,6 @@ public final class CommonInitializer implements ModInitializer {
         // Prepare and run the mod.
         var mod = RelicsMod.instance();
         mod.addSidedFeature(Derelicts.class);
-        mod.addSidedFeature(Elixirs.class);
         mod.addSidedFeature(Relics.class);
         mod.run(Side.Common);
     }