+++ src/main/java/svenhjol/charmony/relics/common/features/elixirs/Constants.java
@@ -1,5 +1,0 @@
-package svenhjol.charmony.relics.common.features.elixirs;
-
-public final class Constants {
-    public static final String ELIXIR_LOOT = "elixir";
-}